AirAware/requirements.txt
Flask>=2.3.0
requests>=2.28.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
gunicorn>=21.2.0